fixed label refinment code

+++ data/floatingobjects.py
@@ -9,7 +9,7 @@ import numpy as np
 import pandas as pd
 from data.utils import get_window, read_tif_image, pad, line_is_closed, \
     split_line_gdf_into_segments, remove_lines_outside_bounds
-from data.label_refinement import refine_masks
+from data.label_refinement import refine_masks_iterative
 
 L1CBANDS = ["B1", "B2", "B3", "B4", "B5", "B6", "B7", "B8", "B8A", "B9", "B10", "B11", "B12"]
 L2ABANDS = ["B1", "B2", "B3", "B4", "B5", "B6", "B7", "B8", "B8A", "B9", "B11", "B12"]
@@ -133,18 +133,15 @@ class FloatingSeaObjectRegionDataset(torch.utils.data.Dataset):
             dtype="uint8"
         )
         if refine_labels:
-            self.maskfile = os.path.join(maskpath, f"{region}_refined.tif")
+            self.maskfile = os.path.join(maskpath, "refined", f"{region}.tif")
         else:
-            self.maskfile = os.path.join(maskpath,f"{region}.tif")
+            self.maskfile = os.path.join(maskpath, "original", f"{region}.tif")
+        os.makedirs(os.path.dirname(self.maskfile), exist_ok=True)
 
-        if not os.path.exists(self.maskfile):
+        if not os.path.exists(self.maskfile) and not self.refine_labels:
+            print("label refinement")
             mask = features.rasterize(self.rasterize_geometries, all_touched=True,
                                       transform=self.geotransform, out_shape=(self.height, self.width))
-            if refine_labels:
-                print(f"load image {self.imagefile}")
-                image, _ = read_tif_image(self.imagefile)
-                print(f"refining labels")
-                mask = refine_masks(image, mask)
 
             with rio.open(self.maskfile, "w", **profile) as dst:
                 dst.write(mask[None])
@@ -204,10 +201,17 @@ class FloatingSeaObjectRegionDataset(torch.utils.data.Dataset):
         window = get_window(line, output_size=self.output_size, transform=self.imagemeta["transform"])
 
         image, win_transform = read_tif_image(self.imagefile, window)
-        mask, win_transform = read_tif_image(self.maskfile, window)
+
+        #mask, win_transform = read_tif_image(self.maskfile, window)
+        with rio.open(self.maskfile, "r") as src:
+            if src.count > 1: # if multiple masks available (i.e., refined labels generated with different parameters in each band)
+                # pick a band randomly
+                mask = src.read(np.random.randint(1, src.count + 1), window=window)
+            else:
+                mask = src.read(1, window=window)
 
         # pad image if at the border
-        image, mask = pad(image, mask[0], self.output_size)
+        image, mask = pad(image, mask, self.output_size)
 
         # to float
         image, mask = image.astype(float), mask.astype(float)
@@ -222,6 +226,7 @@ class FloatingSeaObjectRegionDataset(torch.utils.data.Dataset):
 
         return image, mask, id
 
+    """
     def __getitem__OLD(self, index):
         if hasattr(self, 'npyfolder') and self.item_in_cache(index):
             image, mask, id = self.get_item_from_cache(index)
@@ -265,12 +270,8 @@ class FloatingSeaObjectRegionDataset(torch.utils.data.Dataset):
 
         image = np.nan_to_num(image)
 
-        if self.refine_labels:
-            if len(np.unique(mask)) > 1: # only if labels are present to be refined
-                mask = refine_masks(image, mask)
-
         return image, mask, id
-
+    """
 
 class FloatingSeaObjectDataset(torch.utils.data.ConcatDataset):
     def __init__(self, root, fold="train", **kwargs):
